User Feedback
Addition of a functionally global message box to provide the user with feedback as they interact. Mostly used for safe error handling (fewer unwraps where failure is possible), but also used for positive feedback on successful actions.
Also modified the controller functions to make the order of operations actually work. Now the item list will wait its turn to update.
Also restructured main to provide the inventory controller as a context rather than props. May not be the most idiomatic approach, but it is the lazier one.