Dev Tab Improvements
Added a search bar to the dev tab and support for selecting items via id. Makes it possible to change an item's name without deleting and remaking it. That and some other modifications make it far easier to change item names or categories without loosing stock.
Also added desired stock to the item lists and brought the shopping list tab up to par with home.