Single line commits are cool and fun
While using the app for realsies I noticed after adding an item in the dev tab the search bars would not update. Turns out I'm a goober and didn't process the fact that adding an item would also mean needing to rebuild the item id map.
Also, using 5ch for the width of the stock columns ensures that on all (of the two I have available to test) screen sizes, there is room for up to triple digit values.
I'd like to take the fact that this is the kind of commit I'm doing as a sign this project is nearly complete. Given that my family has already implemented the project into our lives, I both hope that is true and know it probably isn't. But I suppose that's also the fun of it.
If this project goes well, it will be a good opportunity to experiment with sharing binaries. Given that I'm writing this from my Linux PC to serve from my Windows one, it would be cool to not have put quite as much effort into rebuilding for every minor commit.