Message Box Time
Added an attribute to the message boxes for the time they were created. This ensures that even if the message is duplicated, it will be displayed a second time. In other words, the decision to display a duplicate message a second time can now be handled via manual logic rather than (I think) Yew's optimizations. Prior to this change, duplicate messages would not be displayed. In normal use, this meant repeated consuming one item at a time provided no feedback - which is less than ideal.