General Tracking
Attempting to brain-wrangle the idea of generally tracking an item as "Good" "Low" and "Out" using stock 2, 1, and 0 with desired of 1 was painful. So no more.
Greatest downside is that this update modifies the base Item model, so the database requires the following sql command:
"UPDATE items SET track_general = false"
... to initialize the general tracking value of every item. Doable, but annoying.