Change Logging
Added logging to files for changes to database and api warnings. Database logs are passed to website on request (not on load) and displayed in a parsed and vaguely user-friendly manner.
Api warnings logs keep 4 250kb logs + running, database keeps 9 days + running.
Also improved usability when it comes to accidental additions to the dinner list / grocery bag. Now, 0 counts are properly handled and "No Change" can be selected for generally tracked items.
Sometimes, it can be difficult to remember whether or not we subtracted dinner. Now, we can just look at the logs.