Re-Inventory and Desired Stock
This update changes the functionality of stock highlighting. Now, desired stock is intended to be the minimum stock before you get notified. For example, a desired stock of 5 would be classified as low at 4, and out at 3 or less. The major benefit of this change is that it is now possible to properly track items you only want to keep 1 of.
This update also adds a new tab for easier inventorying. Only changes are sent to the database. This tab also allows quickly changing the desired stock of an item, intended for updating generally tracked items.
Updating to this version should be paired with a household re-inventory to ensure desired stocks are accurate.