Significant Rework
Everything is now provided via one application, settings are now accessible,  and logging has seen major changes.

The api now leverages Rocket's file sharing capabilities to serve the web interface as well. The interface was changed to reflect that it is now served from the same location as the database. This is so much better than Windows Internet Services and I am sad I didn't do it from the start. Problem: the web resources need to be copied from their project manually with every update. Another reason to see if I can get distribution up and running.

There is now an external settings file. Gone are the days of hard-coded IP's and ports. This should also make it far easier to keep the logs easily accessible even when working directory is somewhere esoteric - just provide an absolute path to somewhere nicer.

Logging has been completely redone. Given that logging is an important part of the web interface, I saw little reason to continue exposing the logger config. That is now programatically handled, and encoded in something resembling json. This has made the log files even less human readable, but that was barely true to begin with, so it's a minor loss.