Windows things
Popping the end of a text file string twice to remove the newline and a character doesn't work on Windows. Instead, we now utilize the built-in function to remove the ending whitespace, then pop the last character. Should have done that from the beginning, lesson learned.
Starting the program using Windows Task Scheduler can be done silently from a preferred directory is actually pretty easy when you're not bad (like I am). This does seem to mess with the relative! macro, however.